senior/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
psycopg2==2.9.6
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.3.0
channels==4.0.0
channels-redis==4.2.0